+++ coverage/phystokens.py
@@ -27,7 +27,7 @@ def phys_tokens(toks):
     """
     last_line = None
     last_lineno = -1
-    last_ttype = None
+    last_ttext = None
     for ttype, ttext, (slineno, scol), (elineno, ecol), ltext in toks:
         if last_lineno != elineno:
             if last_line and last_line.endswith("\\\n"):
@@ -47,9 +47,7 @@ def phys_tokens(toks):
                 # so we need to figure out if the backslash is already in the
                 # string token or not.
                 inject_backslash = True
-                if last_ttype == tokenize.COMMENT:
-                    # Comments like this \
-                    # should never result in a new token.
+                if last_ttext.endswith("\\"):
                     inject_backslash = False
                 elif ttype == token.STRING:
                     if "\n" in ttext and ttext.split('\n', 1)[0][-1] == '\\':
@@ -66,7 +64,8 @@ def phys_tokens(toks):
                         last_line
                         )
             last_line = ltext
-            last_ttype = ttype
+        if ttype not in (tokenize.NEWLINE, tokenize.NL):
+            last_ttext = ttext
         yield ttype, ttext, (slineno, scol), (elineno, ecol), ltext
         last_lineno = elineno