+++ CHANGES.rst
@@ -33,6 +33,9 @@ Unreleased
 - `debug=plugin` didn't properly support configuration or dynamic context
   plugins, but now it does, closing `issue 834`_.
 
+- Added a JSON report `issue 720`_.
+
+.. _issue 720: https://github.com/nedbat/coveragepy/issues/720
 .. _issue 822: https://github.com/nedbat/coveragepy/issues/822
 .. _issue 834: https://github.com/nedbat/coveragepy/issues/834
 .. _issue 829: https://github.com/nedbat/coveragepy/issues/829