+++ CHANGES.rst
@@ -17,8 +17,10 @@ development at the same time, like 4.5.x and 5.0.
     ..  Version 7.8.1 --- 2021-07-27
     ..  ----------------------------
 
-Unreleased
-----------
+.. _changes_50a7:
+
+Version 5.0a7 --- 2019-09-21
+----------------------------
 
 - Data can now be "reported" in JSON format, for programmatic use, as requested
   in `issue 720`_.  The new ``coverage json`` command writes raw and summarized